Update external/libyuv to r1837
Update external/libyuv to version r1837 (d53f1bee) from upstream.

This brings in some new functions that will be used by libavif.

Also update README.version with local modifications. The files/
subdirectory is a pristine copy of the upstream checkout except
for the local modifications that are listed in README.version.

Bug: b/241008246
Bug: b/228492909
Test: Builds. Media and Camera CTS tests pass.
Change-Id: I1bd041e475666e13fafd8c3792d7142a022d435a