external/libyuv: Move files/fuzz to top level directory
fuzz subdirectory is not part of upstream libyuv and is specific to
android's copy of libyuv. So by moving it to the top level
external/libyuv directory, we can make sure that "files" subdirectory
contains a pristine copy of upstream libyuv.

It also makes updating libyuv from upstream easier.

Bug: 241008246
Test: libyuv_mjpeg_dec_fuzz target builds successfully

Merged-In: Iab37b2c9365725362443408f5fc493730fafcd86
Change-Id: I8684819a8069c5cd0f2b8a36824ee424cf086130